# frontend/client/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:38090","description":"Generated server url"}],"tags":[{"name":"Geozone Controller","description":"API для работы с геозонами"},{"name":"Device Controller","description":"API для работы с устройствами (транспортными средствами)"}],"paths":{"/api/devices/{id}":{"get":{"tags":["Device Controller"],"summary":"Получить устройство по ID","description":"Возвращает устройство по его уникальному идентификатору","operationId":"getDeviceById","parameters":[{"name":"id","in":"path","description":"Уникальный идентификатор устройства","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Устройство не найдено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}},"200":{"description":"Устройство найдено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}}}},"put":{"tags":["Device Controller"],"summary":"Обновить устройство","description":"Обновляет существующее устройство на основе переданных данных","operationId":"updateDevice","parameters":[{"name":"id","in":"path","description":"Уникальный идентификатор устройства","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}},"required":true},"responses":{"404":{"description":"Устройство не найдено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}},"409":{"description":"Устройство с таким deviceId уже существует","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}},"200":{"description":"Устройство успешно обновлено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}}}},"delete":{"tags":["Device Controller"],"summary":"Удалить устройство","description":"Удаляет устройство по его уникальному идентификатору","operationId":"deleteDevice","parameters":[{"name":"id","in":"path","description":"Уникальный идентификатор устройства","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Устройство не найдено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}},"204":{"description":"Устройство успешно удалено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}}}}},"/api/geozones":{"post":{"tags":["Geozone Controller"],"summary":"Создать новую геозону","description":"Создает новую геозону на основе переданных данных","operationId":"createGeozone","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeozoneDto"}}},"required":true},"responses":{"201":{"description":"Геозона успешно создана","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeozoneDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeozoneDto"}}}}}}},"/api/geozones/rectangle":{"post":{"tags":["Geozone Controller"],"summary":"Получить геозоны в прямоугольной области","description":"Возвращает список геозон, находящихся в заданной прямоугольной области","operationId":"getGeozonesInRectangle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RectangleRequest"}}},"required":true},"responses":{"200":{"description":"Список геозон успешно получен","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GeozoneDto"}}}}}}}},"/api/geozones/polygon/s2":{"post":{"tags":["Geozone Controller"],"summary":"Преобразовать полигон в S2 ячейки","description":"Принимает список координат полигона и возвращает список S2 ячеек с границами для отрисовки","operationId":"convertPolygonToS2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PolygonRequest"}}},"required":true},"responses":{"400":{"description":"Некорректные координаты","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/S2ConversionResult"}}}}},"200":{"description":"Успешно преобразовано","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/S2ConversionResult"}}}}}}}},"/api/devices":{"get":{"tags":["Device Controller"],"summary":"Получить список устройств","description":"Возвращает список устройств с поддержкой фильтрации и пагинации","operationId":"getDevices","parameters":[{"name":"name","in":"query","description":"Фильтр по названию (частичное совпадение)","required":false,"schema":{"type":"string"}},{"name":"registrationNumber","in":"query","description":"Фильтр по регистрационному номеру","required":false,"schema":{"type":"string"}},{"name":"typeId","in":"query","description":"Фильтр по типу устройства","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"departmentId","in":"query","description":"Фильтр по ID подразделения","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","description":"Номер страницы (по умолчанию 0)","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","description":"Размер страницы (по умолчанию 20)","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"sort","in":"query","description":"Сортировка (например: createdAt,id или -createdAt для убывания)","required":false,"schema":{"type":"string","default":"createdAt,desc"}}],"responses":{"200":{"description":"Список устройств успешно получен","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeviceDto"}}}}}}},"post":{"tags":["Device Controller"],"summary":"Создать новое устройство","description":"Создает новое устройство на основе переданных данных","operationId":"createDevice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}},"required":true},"responses":{"201":{"description":"Устройство успешно создано","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}},"409":{"description":"Устройство с таким deviceId уже существует","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}}}}},"/api/geozones/{id}":{"get":{"tags":["Geozone Controller"],"summary":"Получить геозону по ID","description":"Возвращает геозону по её уникальному идентификатору","operationId":"getGeozoneById","parameters":[{"name":"id","in":"path","description":"Уникальный идентификатор геозоны","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Геозона найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeozoneDto"}}}},"404":{"description":"Геозона не найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeozoneDto"}}}}}}}},"components":{"schemas":{"DeviceDto":{"required":["deviceId","name","registrationNumber","typeId"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"registrationNumber":{"type":"string"},"deviceId":{"type":"string"},"typeId":{"type":"integer","format":"int32"},"departmentId":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"}},"description":"Данные для создания устройства"},"GeozoneDto":{"required":["coordinates","isActive","lat","lon","name","s2Key","type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string"},"coordinates":{"type":"array","items":{"type":"array","items":{"type":"number","format":"double"}}},"isActive":{"type":"boolean"},"s2Key":{"type":"integer","format":"int64"},"lat":{"type":"number","format":"double"},"lon":{"type":"number","format":"double"}},"description":"Данные для создания геозоны"},"RectangleRequest":{"required":["bottomRightLat","bottomRightLon","topLeftLat","topLeftLon"],"type":"object","properties":{"topLeftLat":{"type":"number","description":"Широта левой верхней точки","format":"double","example":55.7558},"topLeftLon":{"type":"number","description":"Долгота левой верхней точки","format":"double","example":37.6173},"bottomRightLat":{"type":"number","description":"Широта правой нижней точки","format":"double","example":55.7512},"bottomRightLon":{"type":"number","description":"Долгота правой нижней точки","format":"double","example":37.6231}},"description":"Параметры прямоугольной области"},"PolygonRequest":{"required":["coordinates"],"type":"object","properties":{"coordinates":{"type":"array","description":"Список координат полигона, каждая точка - [широта, долгота]","example":[[55.751244,37.618423],[55.755244,37.622423],[55.758244,37.615423]],"items":{"type":"array","description":"Список координат полигона, каждая точка - [широта, долгота]","example":[[55.751244,37.618423],[55.755244,37.622423],[55.758244,37.615423]],"items":{"type":"number","description":"Список координат полигона, каждая точка - [широта, долгота]","format":"double"}}},"maxLevel":{"type":"integer","description":"Максимальный уровень S2 (по умолчанию 24)","format":"int32","example":24}},"description":"Запрос с координатами полигона и уровнем S2"},"S2ConversionResult":{"required":["level","polygon","s2CellId"],"type":"object","properties":{"s2CellId":{"type":"integer","description":"Номер S2 ячейки","format":"int64"},"level":{"type":"integer","description":"Уровень S2 ячейки (0-30)","format":"int32","example":24},"polygon":{"type":"array","description":"Полигон ячейки","example":[[55.7558,55.7558],[55.7558,55.7558]],"items":{"type":"array","description":"Полигон ячейки","example":[[55.7558,55.7558],[55.7558,55.7558]],"items":{"type":"number","description":"Полигон ячейки","format":"double"}}}},"description":"Результат преобразования зоны в S2 ячейки"},"Unit":{"type":"object"}}}}